Refactored code using Clean Code principles – smaller functions, no duplication
