+++ src/components/common/DownloadBalanceSheetButton/DownloadBalanceSheetButton.tsx
@@ -1,9 +1,10 @@
 import { useNavigation } from '@react-navigation/native';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import React from 'react';
-import { TouchableOpacity, Text, StyleSheet, View } from 'react-native';
+import { TouchableOpacity, Text, View } from 'react-native';
+import styles from './DownloadBalanceSheetButton.styles';
 import Icon from 'react-native-vector-icons/Feather';
-import { RootStackParamList } from '../../types/Types';
+import { RootStackParamList } from '../../../types/Types';
 
 const DownloadBalanceSheetButton: React.FC = () => {
   const navigation =
@@ -27,30 +28,5 @@ const DownloadBalanceSheetButton: React.FC = () => {
     </TouchableOpacity>
   );
 };
-
-const styles = StyleSheet.create({
-  button: {
-    backgroundColor: '#f7931e',
-    borderRadius: 30,
-    paddingVertical: 10,
-    paddingHorizontal: 10,
-    alignSelf: 'flex-end',
-    elevation: 4,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.3,
-    shadowRadius: 3,
-  },
-  innerContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#fff',
-    fontWeight: '600',
-    marginLeft: 8,
-    fontSize: 14,
-  },
-});
 
 export default DownloadBalanceSheetButton;
