+++ src/components/common/InputField/InputField.styles.ts
@@ -1,0 +1,15 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  inputContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    backgroundColor: '#f4f4f4',
+    borderRadius: 10,
+    marginBottom: 10,
+    paddingHorizontal: 10,
+  },
+  icon: { marginRight: 8 },
+  input: { flex: 1, paddingVertical: 16 },
+  error: { color: 'red', fontSize: 12, marginBottom: 5, marginLeft: 5 },
+});
