+++ src/components/common/LoaderOverlay/LoaderOverlay.styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  overlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0,0,0,0.4)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
+
