+++ src/components/common/ProductInfo/ProductInfo.styles.ts
@@ -1,0 +1,23 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    marginBottom: 8,
+    padding: 4,
+    borderRadius: 8,
+  },
+  text: {
+    fontSize: 14,
+    color: '#333',
+    marginLeft: 5,
+  },
+  highlight: {
+    fontWeight: 'bold',
+    color: '#FFA500',
+    },
+  statusIcon: {
+    marginLeft: 10,
+  },
+});
