+++ src/components/common/SubmitButton/SubmitButton.styles.ts
@@ -1,0 +1,11 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  button: {
+    backgroundColor: '#FFA500',
+    padding: 12,
+    borderRadius: 30,
+    marginTop: 10,
+  },
+  buttonText: { color: 'white', textAlign: 'center', fontWeight: 'bold' },
+});
