+++ src/components/popup/SuccessPopup/SuccessPopup.styles.ts
@@ -1,0 +1,32 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  overlay: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: 'rgba(0,0,0,0.7)',
+  },
+  container: {
+    backgroundColor: 'white',
+    padding: 20,
+    borderRadius: 12,
+    alignItems: 'center',
+    width: '80%',
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.3,
+    shadowRadius: 6,
+    elevation: 8,
+  },
+  iconContainer: {
+    marginBottom: 15,
+  },
+  message: {
+    fontSize: 18,
+    textAlign: 'center',
+    color: '#333',
+    lineHeight: 24,
+    fontWeight: '500',
+  },
+});
