+++ src/components/popup/SuccessPopup/SuccessPopup.tsx
@@ -1,5 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { Modal, View, Text, StyleSheet } from 'react-native';
+import { Modal, View, Text } from 'react-native';
+import styles from './SuccessPopup.styles';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
 let showSuccess: ((message: string) => void) | null = null;
@@ -38,34 +39,3 @@ export const showSuccessPopup = (msg: string) => {
   if (showSuccess) showSuccess(msg);
   else console.warn('SuccessPopup not mounted in App.tsx');
 };
-
-const styles = StyleSheet.create({
-  overlay: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0,0,0,0.7)',
-  },
-  container: {
-    backgroundColor: 'white',
-    padding: 20,
-    borderRadius: 12,
-    alignItems: 'center',
-    width: '80%',
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 4 },
-    shadowOpacity: 0.3,
-    shadowRadius: 6,
-    elevation: 8,
-  },
-  iconContainer: {
-    marginBottom: 15,
-  },
-  message: {
-    fontSize: 18,
-    textAlign: 'center',
-    color: '#333',
-    lineHeight: 24,
-    fontWeight: '500',
-  },
-});
