+++ src/components/user/TransactionCard/TransactionCard.styles.ts
@@ -1,0 +1,49 @@
+import { StyleSheet } from "react-native";
+
+export default StyleSheet.create({
+  transactionCard: {
+    backgroundColor: '#ffffff',
+    borderRadius: 16,
+    paddingVertical: 18,
+    paddingHorizontal: 20,
+    marginVertical: 10,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.1,
+    shadowRadius: 8,
+    elevation: 5,
+    },
+  productInfoContainer: { flex: 1, marginLeft: 10 },
+  status: { flexDirection: 'row', alignItems: 'center' },
+  productName: { fontSize: 17, fontWeight: '600', color: '#1A1A1A' },
+  timestampText: {
+    fontSize: 13,
+    color: '#666',
+    marginTop: 2,
+    fontWeight: '500',
+  },
+  amountBox: {
+    backgroundColor: '#FFF2E0',
+    borderRadius: 10,
+    paddingVertical: 8,
+    paddingHorizontal: 16,
+    minWidth: 80,
+    alignItems: 'center',
+  },
+  quantityText: {
+    fontSize: 14,
+    color: '#444',
+    marginTop: 4,
+    fontWeight: '500',
+  },
+  amountText: { color: '#F78C1F', fontWeight: '700', fontSize: 15 },
+  added: { color: '#CC8FEC' },
+  removed: { color: '#FF0000' },
+  statusText: { fontWeight: '600', color: '#000' },
+  marginLeft10: { marginLeft: 10 },
+  marginRight10: { marginRight: 10 },
+  editIcon: { color: '#CC8FEC' },
+});
