+++ src/components/user/TransactionList/TransactionList.tsx
@@ -1,15 +1,16 @@
 import React from 'react';
-import { ScrollView, View, Text, Image, StyleSheet } from 'react-native';
-import { CustomerListProps } from '../../types/TransactionTypes';
-import { SECTION_TITLES } from '../../constants/TransactionConstants';
+import { ScrollView, View, Text, Image } from 'react-native';
+import styles from './TransactionList.styles';
+import { CustomerListProps } from '../../../types/TransactionTypes';
+import { SECTION_TITLES } from '../../../constants/TransactionConstants';
 import {
   isRowEmpty,
   shouldIncludeRow,
   getRowDisplayData,
-} from '../../utils/RowUtils';
-import { parseDate, formatOnlyDate } from '../../utils/DateUtils';
-import TransactionCard from './TransactionCard';
-import { InventoryActionType } from '../../services/spreadsheet/google/GoogleSheetService';
+} from '../../../utils/RowUtils';
+import { parseDate, formatOnlyDate } from '../../../utils/DateUtils';
+import TransactionCard from '../TransactionCard/TransactionCard';
+import { InventoryActionType } from '../../../services/spreadsheet/google/GoogleSheetService';
 
 const TransactionList: React.FC<CustomerListProps> = ({
   customers,
@@ -25,7 +26,17 @@ const TransactionList: React.FC<CustomerListProps> = ({
 
   const filteredData = (groupedBySheet[activeTab] || []).filter(row => {
     const rowWithoutSheetName = row.slice(1);
-    return !isRowEmpty(rowWithoutSheetName) && shouldIncludeRow(row, activeTab);
+    const isValidRow =
+      !isRowEmpty(rowWithoutSheetName) && shouldIncludeRow(row, activeTab);
+
+    if (activeTab === 'Inventory') {
+      const qty = Number(row[2]);
+      if (isNaN(qty) || qty <= 0) {
+        return false;
+      }
+    }
+
+    return isValidRow;
   });
 
   const dataWithOriginalIndex = filteredData.map(row => ({
@@ -37,11 +48,13 @@ const TransactionList: React.FC<CustomerListProps> = ({
     (a, b) => parseDate(b.row[3]).getTime() - parseDate(a.row[3]).getTime(),
   );
 
+  console.log(sortedData, 'sortedData');
+
   if (sortedData.length <= 0) {
     return (
       <View style={styles.imageContainer}>
         <Image
-          source={require('../../assets/empty.png')}
+          source={require('../../../assets/empty.png')}
           style={styles.image}
           resizeMode="contain"
         />
@@ -52,7 +65,7 @@ const TransactionList: React.FC<CustomerListProps> = ({
   let lastRenderedDate = '';
 
   return (
-    <ScrollView style={{ flex: 1, padding: 10 }}>
+    <ScrollView style={styles.container}>
       <Text style={styles.sectionTitle}>
         {SECTION_TITLES[activeTab] || activeTab}
       </Text>
@@ -79,8 +92,8 @@ const TransactionList: React.FC<CustomerListProps> = ({
               displayData={displayData}
               activeTab={activeTab}
               onEdit={onEdit}
-              deleteRow={(sheetName: string, index: number) => {
-                deleteRow(sheetName as InventoryActionType, index);
+              deleteRow={(sheetName: string, rowIndex: number) => {
+                deleteRow(sheetName as InventoryActionType, rowIndex);
               }}
               originalIndex={originalIndex}
             />
@@ -90,26 +103,5 @@ const TransactionList: React.FC<CustomerListProps> = ({
     </ScrollView>
   );
 };
-
-const styles = StyleSheet.create({
-  sectionTitle: {
-    fontSize: 20,
-    fontWeight: '700',
-    marginLeft: 8,
-    color: '#1A1A1A',
-  },
-  imageContainer: { alignItems: 'center', marginTop: 100 },
-  image: { width: 300, height: 300 },
-  statusContainer: {
-    backgroundColor: '#FFF2E0',
-    alignSelf: 'center',
-    paddingHorizontal: 16,
-    paddingVertical: 6,
-    fontWeight: '600',
-    color: '#000',
-    borderRadius: 20,
-    marginTop: 16,
-  },
-});
 
 export default TransactionList;
