+++ src/screens/HomeScreen/HomeScreen.tsx
@@ -1,21 +1,13 @@
 import React from 'react';
-import {
-  View,
-  Text,
-  StyleSheet,
-  Image,
-  TouchableOpacity,
-  ScrollView,
-  Dimensions,
-} from 'react-native';
-import Header from '../components/common/Header';
+import { View, Text, Image, TouchableOpacity, ScrollView } from 'react-native';
+import styles from './HomeScreen.styles';
+import Header from '../../components/common/Header/Header';
 import { useDispatch, useSelector } from 'react-redux';
-import { RootState } from '../redux/Store';
-import { setUser } from '../redux/slices/UserSlice';
+import { RootState } from '../../redux/Store';
+import { setUser } from '../../redux/slices/UserSlice';
 import { useNavigation } from '@react-navigation/native';
-import DownloadBalanceSheetButton from '../components/common/DownloadBalanceSheetButton';
-import { useTransactionLogic } from '../hooks/UseTransactionLogic';
-const { height: screenHeight } = Dimensions.get('window');
+import DownloadBalanceSheetButton from '../../components/common/DownloadBalanceSheetButton/DownloadBalanceSheetButton';
+import { useTransactionLogic } from '../../hooks/UseTransactionLogic';
 
 type CardProps = {
   title: string;
@@ -51,25 +43,25 @@ const HomeScreen = () => {
           <Card
             title="Purchase"
             color="#FEC27E"
-            image={require('../assets/purchase.png')}
+            image={require('../../assets/purchase.png')}
             onPress={() => handleCardPress('Purchase')}
           />
           <Card
             title="Sales"
             color="#E2B6FF"
-            image={require('../assets/sales.png')}
+            image={require('../../assets/sales.png')}
             onPress={() => handleCardPress('Sales')}
           />
           <Card
             title="Inventory"
             color="#F4F1ED"
-            image={require('../assets/Ivt.png')}
+            image={require('../../assets/Ivt.png')}
             onPress={() => handleCardPress('Inventory')}
           />
           <Card
             title="Inventory Log"
             color="#FFCB91"
-            image={require('../assets/history.png')}
+            image={require('../../assets/history.png')}
             onPress={() => handleCardPress('Inventory Log')}
           />
         </View>
@@ -91,41 +83,3 @@ const Card: React.FC<CardProps> = ({ title, color, image, onPress }) => (
 );
 
 export default HomeScreen;
-
-const styles = StyleSheet.create({
-  wrapper: {
-    flex: 1,
-    backgroundColor: '#fff',
-    paddingTop: 20,
-  },
-  scrollContent: {
-    minHeight: screenHeight,
-    padding: 16,
-  },
-  cardContainer: {
-    gap: 15,
-    marginTop: 20,
-  },
-  card: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    borderRadius: 12,
-    padding: 6,
-  },
-  cardImage: {
-    width: 100,
-    height: 100,
-    marginRight: 20,
-    resizeMode: 'contain',
-  },
-  cardTextContainer: {
-    flex: 1,
-    alignItems: 'flex-end',
-  },
-  cardText: {
-    fontSize: 20,
-    fontWeight: '600',
-    paddingRight: 20,
-    textAlign: 'right',
-  },
-});
