+++ src/screens/WelcomeScreen/WelcomeScreen.tsx
@@ -1,19 +1,11 @@
 import React from 'react';
-import {
-  View,
-  Text,
-  TouchableOpacity,
-  StyleSheet,
-  Alert,
-  Dimensions,
-} from 'react-native';
+import { View, Text, TouchableOpacity, Alert } from 'react-native';
+import styles from './WelcomeScreen.styles';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { GoogleAuthService } from '../services/spreadsheet/google/GoogleAuthService';
+import { GoogleAuthService } from '../../services/spreadsheet/google/GoogleAuthService';
 import { useNavigation } from '@react-navigation/native';
 import { useDispatch } from 'react-redux';
-import { setAccessToken, setUser } from '../redux/slices/UserSlice';
-
-const { width } = Dimensions.get('window');
+import { setAccessToken, setUser } from '../../redux/slices/UserSlice';
 
 const WelcomeScreen: React.FC = () => {
   const navigation = useNavigation<any>();
@@ -63,60 +55,3 @@ const WelcomeScreen: React.FC = () => {
 };
 
 export default WelcomeScreen;
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#F2F2F2',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 20,
-  },
-  logoText: {
-    fontSize: 32,
-    fontWeight: 'bold',
-    marginBottom: 8,
-  },
-  hisab: {
-    color: '#000',
-    fontWeight: 'bold',
-    letterSpacing: 2,
-  },
-  kitab: {
-    color: '#FC991A',
-    fontSize: 42,
-    fontWeight: 'bold',
-    letterSpacing: 2,
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#888',
-    marginBottom: 100,
-  },
-  logoBox: {
-    alignItems: 'center',
-    marginTop: 100,
-  },
-  subText: {
-    color: '#888',
-    fontSize: 14,
-    marginTop: 8,
-    textAlign: 'center',
-  },
-  button: {
-    position: 'absolute',
-    bottom: 50,
-    backgroundColor: '#FC991A',
-    paddingVertical: 18,
-    paddingHorizontal: 40,
-    borderRadius: 30,
-    width: width * 0.7,
-    alignItems: 'center',
-  },
-  buttonText: {
-    color: '#fff',
-    fontWeight: 'bold',
-
-    fontSize: 16,
-  },
-});
