+++ src/services/spreadsheet/methods/HandleSale.ts
@@ -6,10 +6,12 @@ import { axiosInstance } from "../../config/AxiosInstance";
 import { logInventoryChange } from "./logInventoryChange";
 import { handleError } from "../../../utils/ErrorHandler";
 import { markRowAsUpdated } from "./MarkRowAsUpdated";
-import { showErrorPopup } from "../../../components/popup/ErrorPopup";
-import { showSuccessPopup } from "../../../components/popup/SuccessPopup";
+import { showErrorPopup } from "../../../components/popup/ErrorPopup/ErrorPopup";
+import { showSuccessPopup } from "../../../components/popup/SuccessPopup/SuccessPopup";
 import { SaleData } from "../../../types/Index";
 
+const HEADER_OFFSET = 2;
+const DEFAULT_FLAGS = ["FALSE", "FALSE"];
 
 export async function handleSale(
   spreadsheetId: string | null,
@@ -20,73 +22,33 @@ export async function handleSale(
   setShowModal?: (v: boolean) => void
 ) {
   if (!spreadsheetId || !accessToken) {
-    return  showErrorPopup({ title: "Initialization Error", message: "Spreadsheet ID or access token is missing." });
+    return showErrorPopup({ title: "Initialization Error", message: "Spreadsheet ID or access token is missing." });
   }
 
   try {
     const productName = (data.productName || "").trim();
     const newQty = parseIntSafe(data.quantity);
+
     const inventoryRows = await getSheetData(spreadsheetId, accessToken, "Inventory");
     if (!inventoryRows) return showErrorPopup({ title: "Error", message: "Failed to fetch inventory." });
 
-    const rowIndex = inventoryRows.findIndex(
-      row => normalizeString(row[0]) === normalizeString(productName) && (row[6] || "").toLowerCase() === "false"
-    );
+    const rowIndex = findInventoryRow(inventoryRows, productName);
+    if (rowIndex === -1) return showErrorPopup({ title: "Product Not Found", message: `The product "${productName}" does not exist in inventory.` });
 
-    if (rowIndex === -1) {
-      return showErrorPopup({ title: "Product Not Found", message: `The product "${productName}" does not exist in inventory.` });
-    }
+    const oldQty = editRowIndex !== undefined ? await handleEditSaleMark(spreadsheetId, accessToken, editRowIndex) : 0;
+    const adjustedStock = calculateAdjustedStock(inventoryRows[rowIndex], oldQty, newQty);
 
-    let oldQty = 0;
-    if (editRowIndex !== undefined) {
-      const saleData = await getSheetData(spreadsheetId, accessToken, "Sales");
-      if (!saleData) return showErrorPopup({ title: "Error", message: "Failed to load sale data for editing." });
-      const dataIndex = editRowIndex - 2;
-      const oldRow = saleData[dataIndex];
-      oldQty = oldRow ? parseInt(oldRow[6] || "0", 10) : 0;
-      await markRowAsUpdated(spreadsheetId, accessToken, "Sales", editRowIndex);
-    }
+    if (adjustedStock < 0) return showErrorPopup({ title: "Insufficient Stock", message: `Only ${parseIntSafe(inventoryRows[rowIndex][1])} units available in stock for "${productName}".` });
 
-    const currentStock = parseIntSafe(inventoryRows[rowIndex][1]);
-    const adjustedStock = currentStock + oldQty - newQty;
-
-    if (adjustedStock < 0) {
-      return showErrorPopup({
-        title: "Insufficient Stock",
-        message: `Only ${currentStock} units available in stock for "${productName}".`
-      });
-    }
-
-    const salesData = (await getSheetData(spreadsheetId, accessToken, "Sales")) || [];
-    const transactionId = (salesData.length + 1).toString();
+    const transactionId = await getNextTransactionId(spreadsheetId, accessToken);
     const timestamp = `'${getTimestamp()}`;
 
-    const values = [
-      [
-        transactionId,
-        timestamp,
-        data.name,
-        productName,
-        data.number,
-        data.amount,
-        newQty.toString(),
-        data.message,
-        "FALSE",
-        "FALSE",
-      ],
-    ];
+    const rowValues = buildSaleRow(transactionId, timestamp, data, newQty);
 
-    const success = await appendData(spreadsheetId, accessToken, "Sales", values);
+    const success = await appendData(spreadsheetId, accessToken, "Sales", rowValues);
     if (!success) return showErrorPopup({ title: "Save Failed", message: "Failed to save sale." });
 
-
-    const updatedAt =  `'${getTimestamp()}`;
-    await axiosInstance.put(
-      `/${spreadsheetId}/values/Inventory!B${rowIndex + 2}:C${rowIndex + 2}?valueInputOption=USER_ENTERED`,
-      { values: [[adjustedStock.toString(), updatedAt]] },
-      { headers: { Authorization: `Bearer ${accessToken}` } }
-    );
-
+    await updateInventoryStock(spreadsheetId, accessToken, rowIndex, adjustedStock);
     await logInventoryChange(spreadsheetId, accessToken, productName, newQty - oldQty, "Sales");
 
     showSuccessPopup(editRowIndex !== undefined ? "Sale updated!" : "Sale saved!");
@@ -95,4 +57,57 @@ export async function handleSale(
   } catch (error) {
     handleError("Sale save", error, "An error occurred while saving the sale.");
   }
+}
+
+function findInventoryRow(inventoryRows: string[][], productName: string): number {
+  return inventoryRows.findIndex(
+    row => normalizeString(row[0]) === normalizeString(productName) && (row[6] || "").toLowerCase() === "false"
+  );
+}
+
+async function handleEditSaleMark(spreadsheetId: string, accessToken: string, editRowIndex: number): Promise<number> {
+  const salesData = await getSheetData(spreadsheetId, accessToken, "Sales");
+  if (!salesData) {
+    showErrorPopup({ title: "Error", message: "Failed to load sale data for editing." });
+    return 0;
+  }
+
+  const oldRow = salesData[editRowIndex - HEADER_OFFSET];
+  const oldQty = oldRow ? parseInt(oldRow[6] || "0", 10) : 0;
+  await markRowAsUpdated(spreadsheetId, accessToken, "Sales", editRowIndex);
+
+  return oldQty;
+}
+
+function calculateAdjustedStock(inventoryRow: string[], oldQty: number, newQty: number): number {
+  const currentStock = parseIntSafe(inventoryRow[1]);
+  return currentStock + oldQty - newQty;
+}
+
+async function getNextTransactionId(spreadsheetId: string, accessToken: string): Promise<string> {
+  const salesData = (await getSheetData(spreadsheetId, accessToken, "Sales")) || [];
+  return (salesData.length + 1).toString();
+}
+
+function buildSaleRow(transactionId: string, timestamp: string, data: SaleData, newQty: number): string[][] {
+  return [[
+    transactionId,
+    timestamp,
+    data.name,
+    data.productName,
+    data.number,
+    data.amount,
+    newQty.toString(),
+    data.message,
+    ...DEFAULT_FLAGS
+  ]];
+}
+
+async function updateInventoryStock(spreadsheetId: string, accessToken: string, rowIndex: number, adjustedStock: number) {
+  const updatedAt = `'${getTimestamp()}`;
+  await axiosInstance.put(
+    `/${spreadsheetId}/values/Inventory!B${rowIndex + HEADER_OFFSET}:C${rowIndex + HEADER_OFFSET}?valueInputOption=USER_ENTERED`,
+    { values: [[adjustedStock.toString(), updatedAt]] },
+    { headers: { Authorization: `Bearer ${accessToken}` } }
+  );
 }
