+++ src/services/spreadsheet/methods/UpdateInventoryStock.ts
@@ -5,9 +5,11 @@ import { logInventoryChange } from "./logInventoryChange";
 import { handleError } from "../../../utils/ErrorHandler";
 import { updateRow } from "./UpdateRow";
 import { appendData } from "./AppendData";
-import { showErrorPopup } from "../../../components/popup/ErrorPopup";
-import { showSuccessPopup } from "../../../components/popup/SuccessPopup";
+import { showErrorPopup } from "../../../components/popup/ErrorPopup/ErrorPopup";
+import { showSuccessPopup } from "../../../components/popup/SuccessPopup/SuccessPopup";
 import { InventoryData } from "../../../types/Index";
+
+const HEADER_OFFSET = 2;
 
 export async function updateInventoryStock(
   spreadsheetId: string | null,
@@ -16,13 +18,14 @@ export async function updateInventoryStock(
   editRowIndex?: number,
   fetchCustomerData?: () => void,
   setShowModal?: (v: boolean) => void,
-  skipLog?: boolean,
+  skipLog: boolean = false,
   isMyProduct: boolean = true
 ): Promise<void> {
   if (!spreadsheetId || !accessToken) {
     return showErrorPopup({ title: "Initialization Error", message: "Sheet not initialized" });
   }
-  const updatedAt =`'${getTimestamp()}`;
+
+  const updatedAt = `'${getTimestamp()}`;
   const productName = (data.productName || "").trim();
   const quantity = parseIntSafe(data.quantity);
   const unit = data.unit || "pcs";
@@ -35,14 +38,11 @@ export async function updateInventoryStock(
     const inventoryData = await getSheetData(spreadsheetId, accessToken, "Inventory");
     if (!inventoryData) return showErrorPopup({ title: "Error", message: "Failed to load inventory data" });
 
-    const { rowIndex, existingRow } = findExistingInventoryRow(inventoryData, productName, editRowIndex);
-       let oldQty = 0;
-    let newQty = quantity;
+    const { rowIndex, existingRow } = findInventoryRow(inventoryData, productName, editRowIndex);
 
+    const oldQty = existingRow ? parseIntSafe(existingRow[1]) : 0;
 
     if (existingRow) {
-        oldQty = parseIntSafe(existingRow[1]);
-
       await handleExistingInventoryRow(
         spreadsheetId,
         accessToken,
@@ -55,33 +55,15 @@ export async function updateInventoryStock(
         updatedAt
       );
     } else {
-      await appendInventoryRow(
-        spreadsheetId,
-        accessToken,
-        productName,
-        quantity,
-        data.purchasingPrice,
-        unit,
-        updatedAt,
-        isMyProduct
-      );
+      await appendInventoryRow(spreadsheetId, accessToken, productName, quantity, data.purchasingPrice, unit, updatedAt, isMyProduct);
     }
-if (!skipLog) {
-      const changeQty = editRowIndex !== undefined && existingRow
-        ? newQty - oldQty
-        : newQty;
 
-      await logInventoryChange(
-        spreadsheetId,
-        accessToken,
-        productName,
-        changeQty,
-        "Inventory"
-      );
+    if (!skipLog) {
+      const changeQty = editRowIndex !== undefined && existingRow ? quantity - oldQty : quantity;
+      await logInventoryChange(spreadsheetId, accessToken, productName, changeQty, "Inventory");
     }
-    if (isMyProduct) {
-      showSuccessPopup("Inventory updated!");
-    }
+
+    if (isMyProduct) showSuccessPopup("Inventory updated!");
     setShowModal?.(false);
     fetchCustomerData?.();
   } catch (error) {
@@ -89,24 +71,21 @@ if (!skipLog) {
   }
 }
 
-function findExistingInventoryRow(
+
+function findInventoryRow(
   inventoryData: string[][],
   productName: string,
   editRowIndex?: number
 ): { rowIndex: number; existingRow?: string[] } {
-  const normalized = normalizeString(productName);
-  let rowIndex = -1;
-  let existingRow: string[] | undefined;
+  const normalizedName = normalizeString(productName);
 
-  if (editRowIndex) {
-    rowIndex = inventoryData.findIndex((_, idx) => idx + 2 === editRowIndex);
-    existingRow = rowIndex !== -1 ? inventoryData[rowIndex] : undefined;
-  } else {
-    rowIndex = inventoryData.findIndex(row => normalizeString(row[0]) === normalized);
-    existingRow = rowIndex !== -1 ? inventoryData[rowIndex] : undefined;
+  if (editRowIndex !== undefined) {
+    const rowIndex = inventoryData.findIndex((_, idx) => idx + HEADER_OFFSET === editRowIndex);
+    return { rowIndex, existingRow: rowIndex !== -1 ? inventoryData[rowIndex] : undefined };
   }
 
-  return { rowIndex, existingRow };
+  const rowIndex = inventoryData.findIndex(row => normalizeString(row[0]) === normalizedName);
+  return { rowIndex, existingRow: rowIndex !== -1 ? inventoryData[rowIndex] : undefined };
 }
 
 async function handleExistingInventoryRow(
@@ -124,7 +103,7 @@ async function handleExistingInventoryRow(
   const newStock = Math.max(0, currentStock + quantity);
 
   if (existingRow[7] === "FALSE") {
-    await updateRow(spreadsheetId, accessToken, "Inventory", rowIndex + 2, [
+    await updateRow(spreadsheetId, accessToken, "Inventory", rowIndex + HEADER_OFFSET, [
       existingRow[0] || productName,
       newStock.toString(),
       updatedAt,
@@ -137,32 +116,21 @@ async function handleExistingInventoryRow(
     return;
   }
 
-  const hasChanges =
-    normalizeString(productName) !== normalizeString(existingRow[0]) ||
+  const hasChanges = normalizeString(productName) !== normalizeString(existingRow[0]) ||
     newStock !== currentStock ||
     purchasingPrice !== existingRow[4] ||
     unit !== existingRow[5];
 
   if (hasChanges) {
-    await updateRow(spreadsheetId, accessToken, "Inventory", rowIndex + 2, [
+    await updateRow(spreadsheetId, accessToken, "Inventory", rowIndex + HEADER_OFFSET, [
       ...existingRow.slice(0, 6),
       "TRUE",
       existingRow[7] || "FALSE",
     ]);
 
-    const newRow = [
-      productName,
-      quantity.toString(),
-      updatedAt,
-      "FALSE",
-      purchasingPrice,
-      unit,
-      "FALSE",
-      "TRUE",
-    ];
+    const newRow = [productName, quantity.toString(), updatedAt, "FALSE", purchasingPrice, unit, "FALSE", "TRUE"];
     await appendData(spreadsheetId, accessToken, "Inventory", [newRow]);
   }
-
 }
 
 async function appendInventoryRow(
@@ -175,15 +143,6 @@ async function appendInventoryRow(
   updatedAt: string,
   isMyProduct: boolean
 ) {
-  const newRowData = [
-    productName,
-    quantity.toString(),
-    updatedAt,
-    "FALSE",
-    purchasingPrice,
-    unit,
-    "FALSE",
-    isMyProduct ? "TRUE" : "FALSE",
-  ];
+  const newRowData = [productName, quantity.toString(), updatedAt, "FALSE", purchasingPrice, unit, "FALSE", isMyProduct ? "TRUE" : "FALSE"];
   await appendData(spreadsheetId, accessToken, "Inventory", [newRowData]);
 }
