+++ src/utils/ErrorHandler.ts
@@ -1,4 +1,4 @@
-import { showErrorPopup } from "../components/popup/ErrorPopup";
+import { showErrorPopup } from "../components/popup/ErrorPopup/ErrorPopup";
 
 export function handleError(context: string, error: unknown, userMessage = "An error occurred") {
   console.error(`${context} - error:`, error);
